fix(Runner): update strike rule to track both global and per-task excs
This will cause a slight change in behavior when there is an
intermittent failure in only one of the user's tasks, perhaps do
to an incorrect setting or attribute error on a variable.
The behavioral change will cause failures to cascade into a Halt faster
than they did before, if they happened at all before due to global reset.